requirements.txt
flask
flask_cors
pillow
numpy
scipy
traceback2
gunicorn
pymupdf
requests
bcrypt
pytesseract